requirements.txt
# Versiones compatibles con Python 3.13
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.26.0
pandas>=2.1.0
python-dotenv>=1.0.0
openai>=1.3.0
python-dateutil>=2.8.2
requests>=2.31.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
beautifulsoup4>=4.12.2
lxml>=4.9.3

# PyTorch con soporte para CPU (versión estable más reciente)
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

transformers>=4.36.0
nltk>=3.8.1

tqdm>=4.66.1
sentencepiece>=0.1.99
